+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Tests should now leverage default CloudHarvestStack configurations instead of internally managed `tests/docker-compose.yaml`
 - `task_chain_from_dict()` now merges `**kwargs` with the task chain configuration on class instantiation
 - Removed the `HeartBeat` class. Moved `silos.py` into the root package directory
+- Added `HarvestRecord.assign_value_to_key()`
 
 # 0.4.2
 - Added `BaseHarvestTaskChain` and `HarvestRecordUpdateTask` to upload data collected from other sources to a persistent silo